refactor: 기능에 맞는 파일 분리
